removed useFirebase hook and added firebase-utils instead
